requirements.txt
pathlib2
PyYAML
